fix: v662 property sync data field types and conditional move actor delta

+++ crates/protocol/src/version/v662/types/property_sync_data.rs
@@ -11,14 +11,14 @@ pub struct PropertySyncData {
 pub struct IntEntry {
     #[endianness(var)]
     pub property_index: u32,
-    #[endianness(le)]
-    pub data: f32,
+    #[endianness(var)]
+    pub data: i32,
 }
 
 #[derive(ProtoCodec, Clone, Debug)]
 pub struct FloatEntry {
     #[endianness(var)]
     pub property_index: u32,
-    #[endianness(var)]
-    pub data: i32,
+    #[endianness(le)]
+    pub data: f32,
 }
